requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pandas==2.2.2
numpy==2.0.1
scipy==1.13.1
scikit-learn==1.5.1
streamlit==1.36.0
requests==2.32.3
openai==1.35.10